+++ lib/src/widgets/solid_file_browser_content.dart
@@ -26,8 +26,8 @@ library;
 import 'package:flutter/material.dart';
 
 import 'package:solidui/src/models/file_item.dart';
-import 'package:solidui/src/widgets/directory_list.dart';
-import 'package:solidui/src/widgets/file_list.dart';
+import 'package:solidui/src/widgets/solid_file_directory_list.dart';
+import 'package:solidui/src/widgets/solid_file_list.dart';
 
 /// Content widget for the file browser.
