+++ lib/src/widgets/solid_file_empty_directory_view.dart
