+++ lib/src/widgets/solid_file_list.dart
@@ -26,7 +26,7 @@ library;
 import 'package:flutter/material.dart';
 
 import 'package:solidui/src/models/file_item.dart';
-import 'package:solidui/src/widgets/file_list_item.dart';
+import 'package:solidui/src/widgets/solid_file_list_item.dart';
 
 /// A widget that displays a list of files with their metadata and actions.
