+++ lib/src/widgets/solid_file_list_item.dart
