+++ lib/src/widgets/solid_file_path_bar.dart
